requirements.txt
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.9
python-dotenv